% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{sarm1}
\alias{sarm1}
\title{Zhu et al: Effects of SARM1 deficiency on gene expression}
\format{
A DESeqDataSet with 53801 genes and 16 samples.
}
\source{
\href{https://www.ncbi.nlm.nih.gov/sra?term=SRP178253}{Short Read Archive, accession SRP178253}
}
\usage{
data(sarm1)
}
\description{
A DESeqDataSet with raw RNA-seq data published by
\href{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC6446871/}{Zhu et al}. Samples
fall into one of four groups:
\itemize{
\item \code{WT_ctrl}: wildtype genotype, control treatment
\item \code{KO_ctrl}: SARM1-/- genotype, control treatment
\item \code{WT_prion}: wildtype genotype, prion treatment
\item \code{KO_prion}: SARM1-/- genotype, prion treatment
}
}
\keyword{datasets}
