% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{tau}
\alias{tau}
\title{Wang et al: time course of microglia from the rTg4510 mouse model}
\format{
A DESeqDataSet with 53801 genes and 32 samples.
}
\source{
\href{https://www.ncbi.nlm.nih.gov/sra?term=SRP172787}{Short Read Archive, accession SRP172787}
}
\usage{
data(rnai)
}
\description{
A DESeqDataSet with raw RNA-seq data published by
\href{https://www.ncbi.nlm.nih.gov/pubmed/30558641}{Wang et al}.
Samples fall into one of 8 groups, each corresponding to a genotype x time
point combination:
\itemize{
\item \code{WT_2}: microglia isolated from wildtype animals at 2 months of age, N = 4
\item \code{WT_4}: microglia isolated from wildtype animals at 4 months of age, N = 4
\item \code{WT_6}: microglia isolated from wildtype animals at 6 months of age, N = 4
\item \code{WT_8}: microglia isolated from wildtype animals at 8 months of age, N = 4
\item \code{rTg4510_2}: microglia isolated from rTg4510 animals at 2 months of age,
N = 4
\item \code{rTg4510_4}: microglia isolated from rTg4510 animals at 4 months of age,
N = 4
\item \code{rTg4510_6}: microglia isolated from rTg4510 animals at 6 months of age,
N = 4
\item \code{rTg4510_8}: microglia isolated from rTg4510 animals at 8 months of age,
N = 4
}
}
\note{
More information about the rTg4510 mouse model is available on the
\href{https://www.alzforum.org/research-models/rtgtaup301l4510}{Alzforum website}
}
\keyword{datasets}
